changed playing behavior to match animal well
before, pushing any direction played a note -- this is not how it works in animal well. now, only a note is played when the player presses the `A` button, and the pitch can be raised by a semitone when the `B` button is held (this is how the animal flute behaves).